requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
arxiv==1.4.8
langchain==0.0.350
langchain-groq==0.0.1
langchain-google-genai==0.0.8
pydantic==2.5.0
requests==2.31.0
httpx==0.28.1          
gunicorn==21.2.0
